# docs/go-api/swagger.yaml
basePath: /
definitions:
  user_model.User:
    properties:
      id:
        type: string
      name:
        type: string
    required:
    - name
    type: object
host: localhost:3001
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Fiber Swagger Example API
  version: "2.0"
paths:
  /users:
    get:
      consumes:
      - application/json
      description: Lists all users details.
      parameters:
      - description: name
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              items:
                $ref: '#/definitions/user_model.User'
              type: array
            type: array
      summary: Lists all users details.
      tags:
      - Users
    patch:
      consumes:
      - application/json
      description: Update user details
      parameters:
      - description: User data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/user_model.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user_model.User'
      summary: Update user
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Create user details
      parameters:
      - description: User data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/user_model.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user_model.User'
      summary: Create user
      tags:
      - Users
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Display user details
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user_model.User'
      summary: Display user details
      tags:
      - Users
    get:
      consumes:
      - application/json
      description: Display user details
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user_model.User'
      summary: Display user details
      tags:
      - Users
schemes:
- http
swagger: "2.0"